requirements.txt
flask
python-dotenv
requests
Flask
PyJWT
Werkzeug
